requirements.txt
django
djangorestframework
mysqlclient
pillow
djangorestframework-simplejwt
drf-flex-fields
django-filter
django-rest-auth
django-allauth
dj-rest-auth
django-simple-sms
